Use integer instead of bigint for columns that should just need integers

+++ deploy-config/sql/01-catalog-00.sql
@@ -113,8 +113,8 @@ CREATE INDEX idx_slug ON products (slug);
 
 CREATE TYPE product_relation_type AS ENUM ('related', 'upsell', 'cross_sell');
 CREATE TABLE product_relations (
-    source_product_id  BIGINT NOT NULL REFERENCES products(id) ON DELETE CASCADE,
-    target_product_id  BIGINT NOT NULL REFERENCES products(id) ON DELETE CASCADE,
+    source_product_id  INTEGER NOT NULL REFERENCES products(id) ON DELETE CASCADE,
+    target_product_id  INTEGER NOT NULL REFERENCES products(id) ON DELETE CASCADE,
     relation_type      product_relation_type NOT NULL,
     sort_order         INTEGER DEFAULT 0,
     created_at         TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
